requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
selectolax>=0.3.0
trafilatura>=1.6.0
pandas>=2.0.0
openpyxl>=3.1.0

# LLM
openai>=1.0.0  # DeepSeek uses OpenAI-compatible API

# Search
# Using httpx (already in core dependencies) for Google CSE API

# Database
aiosqlite>=0.19.0
psycopg2-binary>=2.9.0  # PostgreSQL driver for Supabase

# CLI
typer>=0.9.0
rich>=13.0.0

# PDF parsing
pymupdf>=1.23.0

# Optional: JS rendering
# playwright>=1.40.0

# Optional: UI
streamlit>=1.29.0

